vlan: add one VLAN in Kconfig as an option
WARNING: wr_switch_defconfig doesn't build, this is fixed in next commit

Unfortunately this commit makes the internal communication between
minic and lan asymmetric: on receive the tag is discarded (and the
longer header falls into the payload) while on transmit it must be
provided by lan.c.

The reason is that on receive we can trim 4 bytes from the payload,
but on send we can't add 4 bytes without a memmove.  The functions
receive pointers to two different header structures, so hopefully
user errors will be signalled by the compiler.

Still, users interact with net.c, which hides vlan completely from
them.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>