Support for host builds.
This commit is not breaking the current status, but adds the
ability to build on the host, to simulate and (mainly) to help
me hack the networking code in a simpler/faster environment.

Missing features:
	- ptp
	- build in i386 (currently x86-64 is hardwired)

Also, this commit requires a change in ppsi, that I'm not going to
commit now: we must get rid of the main function (most of that
stuff is done win wrc_ptp.c, which we must replicate  in a unix-compatible
way).

This is the ppsi change

   --- a/arch-unix/Makefile
   +++ b/arch-unix/Makefile
   @@ -4,7 +4,7 @@ A := arch-$(ARCH)

    CFLAGS += -Itools

   -OBJ-y += $A/unix-startup.o \
   +OBJ-y += \
           $A/main-loop.o \
           $A/unix-io.o \
           $A/unix-conf.o \

Local changes in this commit:
      - adding CONFIG_HOST_PROCESS and the auto-generated opposites:
      CONFIG_LM32 and CONFIG_EMBEDDED_NODE (LM32 && NODE)

      - make some config option depend on !HOST_PROCESS

      - Makefile: some dependencies on HOST_PROCESS and the opposite

      - */*.mk: some dev and shell objects depend on
      CONFIG_EMBEDDED_NODE instead of CONFIG_WR_NODE (i.e.: both a
      node and an lm32)

      - new host/ subdir where some placeholders are placed

      - no sdb support on the host

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>