lib/latency: a latency-test engine, to check WR networks
This adds a tool to check the latency in the network. It is selected
at build time (config) time.  If you select CONFIG_LATENCY_PROBE you
are then asked for a CONFIG_LATENCY_ETHTYPE.

The sender node sends three frames, as broadcast in that ethertype
within the current vlan (or no vlan).  The first frame is at priority
7, the second at priority 6, and the third includes the egress
timestamps for the former ones. Each group has a sequence number
and group identifier.

If the tool is built in wrc.bin, every node acts as a receiver. To
enable a sender issuing the set of frames every 5 seconds, use "ltest 5".
To disable use "ltest 0".

The receiver prints the measured latency, and warns if it receives
frames that are not in the proper order, using the sequence and frame
type fields.

This is a successful run crossing two wr switches:

    latency prio7: 5142.775
    latency prio6: 4998.775
    latency prio7: 4822.773
    latency prio6: 5142.773
    latency prio7: 5142.771
    latency prio6: 5286.771

Later on, we'll make the thing less verbose, and we'll have threshold
to signal a problem if the network went slow. Currently we just print
all collected data.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>