pfilter: accept unicast ARP frames
This commit removes the requirement for ARP frames to be broadcast
in the no-vlan case.  The simplified rule-set of a vlan setup already
accepts ARP unicast, and no harm is done.

Although we do not make ARP  requests so we won't receive unicast
replies, hosts do send unicast requests, to refresh an ARP entry
that is going to expire.

The next commit enacts this feature in software, as a side effect
of supporting pdelay PTP.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>