latency: add syslog support (lost frames count is wrong)
If !CONFIG_SYSLOG, usual verbose output is printed to console.

If CONFIG_SYSLOG, you can send the latency error problems to syslog.
The node sends to syslog only the number of lost frames (FIX THIS)
and the latency if it's more than twice the current running average.
such an average is on 8 samples, so it moves quickly.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>