Kconfig: fix warnings
The new config tools are more picky. We got:

   Kconfig:22:warning: defaults for choice values not supported

and

  warning: (WR_SWITCH) selects UART which has unmet direct dependencies
             (DEVELOPER && EMBEDDED_NODE)

The latter is fixed adding WR_SWITCH in the dependencies. The former
is worked-around by using "IS_XINT" and friends in "choice", and
then "XINT" replicates the "IS_" choice.  I dislike it, because it's
repetition, but we really need CONFIG_PRINTF_XINT and friends in
pp_printf/

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>