Support e2e-only builds
Now the default for configurations that do not select peer-delay
is to not even build peer-delay code. This provides for that, and
adds two configurations for peer-delay operation.

If pdelay is selected, e2e is built as well, but selection is currently
only at build time. This will be fixed later.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>