config: reduce configs/* to the bare minimum
Use "make savedefconfig" and save the resulting short file
(so only what is different from the default is being recorded).
Thus, when adding a new option, there's no need to update configs/*

As a side effect, if we change the default, it applies for everyone (I
think it's good, but it's debatable).

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>