shell: replace strcasecmp with strcmp
when trying a compiler built with a newer newlib, I got a size increase
of 2kB in wrc.elf. This is due du "smarter" strcasecmp, which is
now localized, with wide-character support and so on.

We may implement our own ascii-only strcasecmp, but avoinding case-insensitive
comparisons we save 400 bytes, and the binary size doesn't change changing
newlib in the compiler.

Nobody is expected to use uppercase commands in the wrpc shell, and there
is no technical need to support them, when all documentation hints to
use lower-case anyways.

Change trivially performed by

   sed -i s/strcasecmp/strcmp/ shell/*.c

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>