Fixes on the HO flow operation
Prior to this commit, the holdover operation would only apply the latest "good" position of the DAC. This has been fixed by not disabling irqs altogether and findin a more "proper" way to do this. Now, the tagger is disabled in holdover mode, and we prevent the reset of the softpll caused by a link down event.