make (almost) every call to ertm_proto_cycle atomic
To make libertm thread-safe, calls to ertm_proto_cycle must be completed
in an atomic manner, ensuring that the serial link to the wrpc is used
in an exclusive manner by each client thread.

There is only one exception to this rule, and a crucial one: there are
separate operations for sending config values to the wrpc, and for
committing the sent values. The doublet of the two calls to
set_board_config and commit_board_config, which are called only in one
place and always in sequence, must be atomic in its entirety, which is
most simply achieved by defining a larger critical section in
update_board_config. This is ugly, but the shortest path to fully atomic
operations from the way the code was originally organized.